requirements.txt
pytest==7.2.1
selenium==4.8.2
allure-pytest==2.13.0
pytest-xdist==3.2.0
requests==2.28.1
Faker==17.0.0